seperate the connection to DB to individual file
