+++ config/connection.js
@@ -1,0 +1,13 @@
+const mysql = require('mysql2');
+require('dotenv').config();
+
+const db = mysql.createConnection(
+    {
+      user: process.env.DB_USER,
+      password: process.env.DB_PASSWORD,
+      database: 'company_db',
+    },
+    console.log(`Connected to the company_db database.`),
+  );
+
+module.exports = db;
